Added confirm dialog to remove friend

+++ app/scripts/directives/user.js
@@ -66,18 +66,20 @@ app.directive('removeFriend', function(UserService) {
     },
     link: function(scope, elem, attrs) {
       elem.bind('click', function() {
-        // Make sure target is a number
-        if(angular.isDefined(scope.target)) {
-          elem.prop('disabled', true);
+        if(confirm('Are you sure you want to remove this user as a friend?')) {
+          // Make sure target is a number
+          if(angular.isDefined(scope.target)) {
+            elem.prop('disabled', true);
 
-          // Call remove friend function for User Service
-          UserService.remove_friend(scope.target).then(function(response) {
-            if(response.result) {
-              elem.text('Friend Removed');
-            } else {
-              elem.text('Removal Failed');
-            }
-          });
+            // Call remove friend function for User Service
+            UserService.remove_friend(scope.target).then(function(response) {
+              if(response.result) {
+                elem.text('Friend Removed');
+              } else {
+                elem.text('Removal Failed');
+              }
+            });
+          }
         }
       });
     }
